Use the filter option of clone to avoid cloning everything (#201)
This clones without trees and blob. It gives references to commits and it's possible to fetch the blobs and trees needed on demand.
This reduces the stack-config time by 50%.
I separated the fetch from the checkout but it could be done under the same if, let me know what you prefer.