Spack v0.23 (#204)
Add Spack v0.23 to the list of supported versions.
Support v0.20 is removed.

With the simplifications below, there are now only two locations in the templates where the spack_version is used - most of the changes were removing v0.20 features.

The CLI interface obtained a new CLI flag `--spack_version` which lets the user specify the version of Spack, in case they are using a commit hash and are not on `develop`, e.g.
```
stack-config ... --spack_version=v0.23
```

Internally, the `develop` is no longer passed to the templates - instead only `spack_version` is passed, and `--develop` will set `spack_version` to the next release. This simplifies template logic to deal solely with version numbers.