First class network stack support (#256)
The method for adding MPI and tuning the networking libarary stack has
been simplified.

There are two components that define the network stack.
The first is the new `network` field in `environments.yaml` that
replaces the old `mpi` field.
The user provides a `spec` for the MPI distribution they want (currently
cray-mpich and openmpi are supported), and an optional list of specs for
dependencies (e.g. libfabric):

```yaml
network:
  mpi: <spec>
  specs: [<spec>, ...]
```

The other half is the new `network.yaml` file in the system
configuration, that provides a default list of `specs` for each MPI
distribution, and a `packages.yaml` description of the default /
required variants and options for cray-mpich, openmpi, libfabric, etc.

For example, using this approach, the `+cuda` variant can be set as a
default on a GH200 cluster, and disabled on a CPU-only system.

Based off #210.

---------

Co-authored-by: Alberto Invernizzi <9337627+albestro@users.noreply.github.com>
Co-authored-by: Rocco Meli <r.meli@bluemail.ch>