Reduce the number of Spack environments (#258)
Remove a long-standing work around to an old Spack bug, that required us
to generate a separate Spack environment for every view.
This meant that if an `environments.yaml` recipe contained more than one
view in an environment (e.g. `default` and `develop`), then the
environment would be concretised twice.

Spack 1.0 fixes this issue, so we can .

This has no impact on users or recipes, it is a back end optimization.

Fixes #209