Concretizer duplicates strategy (#269)
Based on top of #268. It's easiest to only look at the unique commit for
this PR:
https://github.com/eth-cscs/stackinator/commit/3ed904aa9fb0b933ebf3116edf1cc2f0fce5a00b.

This is the second change I need to be able to concretize `rccl-tests`
with `llvm-amdgpu` as the compiler. This adds a configuration option to
control the `concretizer:duplicates:strategy` configuration option. For
`rccl-tests` I need to set the option to `full`, instead of the default
`minimal`.

Unfortunately I can't really explain why this is required. The docs
explain a bit what the different options do:
https://spack.readthedocs.io/en/latest/build_settings.html.
```
  duplicates:
    # "none": allows a single node for any package in the DAG.
    # "minimal": allows the duplication of 'build-tools' nodes only
    # (e.g. py-setuptools, cmake etc.)
    # "full" (experimental): allows separation of the entire build-tool stack (e.g. the entire "cmake" subDAG)
```
I've asked for feedback on the spack slack:
https://spackpm.slack.com/archives/C059JUS9T38/p1761907568615009.

In the meantime I'm going to assume that this will be needed. I don't
quite like that this is another ad-hoc spack configuration option
exposed in the config. It'd be nice to just have a `concretizer` section
exposed in `environments.yaml`, but that's a breaking change (`unify` is
not nested under `concretizer`). Alternatively, perhaps it should be
possible to supply a completely separate spack configuration that can be
used to override anything, without having to add support for every
single configuration option explicitly.

---------

Co-authored-by: Alberto Invernizzi <9337627+albestro@users.noreply.github.com>