modules: enforce `modules:default:roots:tcl` to be unset and `arch_folder:false` (#276)
This is a minor bug fix, since it does not apply to any recipe currently
published in https://github.com/eth-cscs/alps-uenv, but it represents a
good sanity check that might help users.

Since the default configuration for spack would be to consider
`arch_folder: true`, if the recipes don't explicitly set it, the modules
get generated with an additional folder level, namely the arch_folder
(e.g. `/user-environment/modules/linux-sles15-neoverse_v2` instead of
simply `/user-environment/modules`).

The problem arises with usage, because `uenv` sets `MODULEPATH` to
`/user-environment/modules`, so the modules are not anymore reachable
simply by name, but they need to be prefixed with the specific
arch_folder (e.g. `linux-sles15-neoverse_v2/cmake` instead of `cmake`).
That does not translates just to a burden for the user that has to
prefix all module names with the arch_folder, but it actually does not
work because modules dependencies in modules themselves are not prefixed
with the `arch_folder`.

Spack implements this functionality by adding to `MODULEPATH` all
arch_folders available on the system.

I think we can live without functionality, but we should anyway enforce
it to catch earlier (at build time) potential problems (which would
otherwise show up just at usage time).