Strip quotes from environment variables when parsing from activate.sh (#277)
This fixes parsing of environment variables that are quoted in
`activate.sh`, such as

```
export NCCL_NET='AWS Libfabric'
```

Before this PR the above would be parsed to the value `'AWS Libfabric'`,
literally, including the quotes. The quotes would be removed by bash,
but when uenv sets environment variables the quotes are retained. In the
particular case of `NCCL_NET` the quoted value is not a valid name for a
plugin name, so NCCL fails to find the plugin.